requirements.txt
openai
sentence_transformers
faiss-cpu
pillow
python-bidi
flask
werkzeug
assemblyai
requests
tensorflow
pandas
tf-keras